Refactor frontend api requests into service file
